add component stats y tabs a pokemonDetails
